server/requirements.txt
aiofiles==0.5.0
alembic==1.5.4
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.1
cryptography==3.3.1
dnspython==2.1.0
docker==4.4.1
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.62.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
orjson==3.4.7
passlib==1.7.4
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.11.8
uvloop==0.14.0
websocket-client==0.57.0
websockets==8.1
